move set_scalar_value functor to utilities/value_to_scalar.hpp, remove test for ListSeries.setValue until cuDF supports scatter for list_scalar
