+++ modules/cudf/test/series/list-tests.ts
@@ -65,6 +65,15 @@ describe('ListSeries', () => {
     }
   });
 
+  // Uncomment this once libcudf supports scatter w/ list_scalar
+  // test('Can set individual values', () => {
+  //   const listOfLists = [[0, 1, 2], [3, 4, 5], [6, 7, 8], [9, null]];
+  //   const col         = Series.new(listOfLists);
+  //   col.setValue(0, [1, 1, 1]);
+  //   expect([...col].map((elt) => [...elt!])).toEqual([[1, 1, 1], [3, 4, 5], [6, 7, 8], [9,
+  //   null]]);
+  // });
+
   test('Can create a List of Lists from Arrow', () => {
     const vec  = listsOfListsOfInt32s([[[0, 1, 2]], [[3, 4, 5], [7, 8, 9]]]);
     const list = vec.getChildAt<ListOfInt32>(0)! as VectorType<ListOfInt32>;
